Design Updates - Back to Basics
There was too much cruft going on with all the styles, and I really
didn't feel like I could make changes without cascading unintended
consequences. So we went back to basics. Less than 100 lines of fairly
straightforward css. There's still some work to be done, but at least
for now it's so much cleaner.

Goodbye minima, thanks for the start but it was time for a change.